requirements.txt
# Core Web
flask==3.1.2
flask-cors==6.0.2
flask-socketio==5.6.0
flask-jwt-extended==4.7.1
flask-mail==0.10.0

# Databases and Messaging
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
redis==7.1.0
neo4j==6.1.0

# ML and Data
numpy==2.4.1
pandas==3.0.0
scipy==1.17.0
scikit-learn==1.8.0
joblib==1.5.3
threadpoolctl==3.6.0
gensim==4.4.0
nltk==3.9.4

# Threat Intel and Security
requests==2.32.5
python-dotenv==1.2.1
mitreattack-python==5.3.0
vulners==3.1.6
cvss==3.6

# Scheduling and Realtime
apscheduler==3.11.2
schedule==1.2.2
python-socketio==5.16.0

# Telemetry
opentelemetry-api==1.41.0
opentelemetry-sdk==1.41.0